+++ src/index.ts
@@ -1,4 +1,3 @@
-export { generateToken } from "./tokens";
-export { default as words } from "./tokens/words.json";
-export { calculateEstimate } from "./calculateEstimate";
-export * from "./Popup";
+export * from "./calculateEstimate/calculateEstimate";
+export * from "./generateToken/generateToken";
+export * from "./Popup/Popup";
